rename test folder
