+++ controllers/api/auth_test.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"go-google-scraper-challenge/initializers"
-	. "go-google-scraper-challenge/test/helpers"
-	. "go-google-scraper-challenge/test/serializers"
+	. "go-google-scraper-challenge/tests/helpers"
+	. "go-google-scraper-challenge/tests/serializers"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
