+++ controllers/session_test.go
@@ -5,7 +5,7 @@ import (
 
 	"go-google-scraper-challenge/controllers"
 	"go-google-scraper-challenge/initializers"
-	. "go-google-scraper-challenge/test/helpers"
+	. "go-google-scraper-challenge/tests/helpers"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
